feat(sql): usable schema exploration, decidable SELECT *, real durations (2.4.0)
From a real session driving these tools against a live instance.

dolibarr_sql_schema was close to unusable for its main purpose. Called with
no argument it returned every column of every table — ~380 KB on an ordinary
instance — which the server then cut at 200 tables in alphabetical order, so
the answer stopped somewhere in the "e"s and llx_facture was simply absent,
with nothing telling the caller. It now returns table NAMES only in that
case, read from information_schema.TABLES: 393 tables in 13 KB here, nothing
truncated, and the columns are one call away on the table that matters.

Called WITH a name it treated it as a prefix, so "llx_facture" answered with
llx_facture, llx_facturedet, llx_facture_extrafields and a dozen more, big
enough to break a client's response limit — which is what "the exact name
fails but a shortened one works" turned out to be. An exact match now wins;
the prefix behaviour is kept for names matching no table.

SELECT * was refused outright. The reasoning held — a third-party table can
hold a key nobody listed — but the cost was paid on every ordinary query.
The star is now decided instead of assumed: the host resolves the real
columns of every table in the statement and the same policy runs over them.
So SELECT * works on business tables, and is refused on llx_user with the
offending column named. Without a resolver, or when resolution yields
nothing, the old refusal stands — "could not find out" must never read as
"nothing to hide".

duration_ms was an int, so every fast query reported 0 and looked like a
timer that was never wired. It is fractional now.

The tool description also states that datef and date_lim_reglement are date
while datec and tms are datetime, which cost the caller a wrong first guess.

+++ src/Sql/SqlExecutionResult.php
@@ -21,7 +21,7 @@ class SqlExecutionResult
         private array $rows,
         private int $rowCount,
         private bool $truncated,
-        private int $durationMs,
+        private float $durationMs,
         private int $bytes,
     ) {
     }
@@ -48,7 +48,13 @@ class SqlExecutionResult
         return $this->truncated;
     }
 
-    public function durationMs(): int
+    /**
+     * Wall-clock duration in milliseconds, kept fractional on purpose.
+     *
+     * Rounded to an int, every fast query reported 0 ms, which reads as a timer
+     * that was never wired rather than as a query that was simply quick.
+     */
+    public function durationMs(): float
     {
         return $this->durationMs;
     }
